% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countymap_v0.1.R
\name{county_level_map}
\alias{county_level_map}
\title{summarize county level MOVESoutput}
\usage{
county_level_map(data, path, DSfile, year, pollutantID, outpath, unit)
}
\arguments{
\item{data}{the data summaryzed by summary_MOVES_data function}

\item{year}{the year you would like to plot}

\item{pollutantID}{pollutant ID NOx 31 VOC 81 PM2.5 110}

\item{path:}{Design sheet path}

\item{DSfile:}{Design sheet file}

\item{hostname:}{SQL sever location}

\item{outpath:}{path of county summary for all pollutant (csv file)}
}
\value{
a county level summary for NOx VOC PM2.5 emissions
}
\description{
This function summarize county level MOVESouput
}
\examples{
county_level_map(data, path, DSfile, year, unit, outpath)

}
