% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieplot_v0.1.R
\name{pieplot}
\alias{pieplot}
\title{summarize county level MOVESoutput}
\usage{
pieplot(data, path, DSfile, year, poll, outpath)
}
\arguments{
\item{data}{the data summaryzed by summary_MOVES_data function}

\item{year}{the year you would like to plot}

\item{path:}{Design sheet path}

\item{DSfile:}{Design sheet file}

\item{hostname:}{SQL sever location}

\item{outpath:}{path of county summary for all pollutant (csv file)}

\item{unit}{the emission unit}
}
\value{
a county level summary for NOx VOC PM2.5 emissions
}
\description{
This function summarize county level MOVESouput
}
\examples{
pieplot(data, path, DSfile, year, pollutantID, outpath)

}
